add test, make execution as a module possible
* allows to use pyminifier without installing
* requires main() to be in __main__
* moved the minify() method to __init__
* added a test which tests if a compressed version of pyminifier gives the same results as the original one..
